+++ src/beicon/core.cljs
@@ -102,6 +102,12 @@
        (.connect ob'))
      ob')))
 
+(defn share
+  "Returns an observable sequence that shares a single
+  subscription to the underlying sequence."
+  [ob]
+  (.share ob))
+
 (defn connect!
   "Connect the connectable observable."
   [ob]